+++ src/code42cli/securitydata/arguments/search.py
@@ -1,20 +1,28 @@
-from code42cli.securitydata.options import ExposureType, OutputFormat
+from code42cli.securitydata.options import ExposureType
 
 
-def add_all_arguments_to_parser(parser):
+def add_arguments_to_parser(parser):
     _add_advanced_query(parser)
     _add_begin_date_arg(parser)
     _add_end_date_arg(parser)
-    _add_output_format_arg(parser)
-    _add_incremental_arg(parser)
     _add_exposure_types_arg(parser)
+
+
+class SearchArguments(object):
+    ADVANCED_QUERY = "advanced_query"
+    BEGIN_DATE = "begin_date"
+    END_DATE = "end_date"
+    EXPOSURE_TYPES = "exposure_types"
+
+    def __iter__(self):
+        return iter([self.ADVANCED_QUERY, self.BEGIN_DATE, self.END_DATE, self.EXPOSURE_TYPES])
 
 
 def _add_advanced_query(parser):
     parser.add_argument(
         "--advanced-query",
         action="store",
-        dest="advanced_query",
+        dest=SearchArguments.ADVANCED_QUERY,
         help="A raw JSON file event query. "
         "Useful for when the provided query parameters do not satisfy your requirements."
         "WARNING: Using advanced queries ignores all other query parameters.",
@@ -26,7 +34,7 @@ def _add_begin_date_arg(parser):
         "-b",
         "--begin",
         action="store",
-        dest="begin_date",
+        dest=SearchArguments.BEGIN_DATE,
         help="The beginning of the date range in which to look for events, "
         "in YYYY-MM-DD UTC format OR a number (number of minutes ago).",
     )
@@ -37,31 +45,9 @@ def _add_end_date_arg(parser):
         "-e",
         "--end",
         action="store",
-        dest="end_date",
+        dest=SearchArguments.END_DATE,
         help="The end of the date range in which to look for events, "
         "in YYYY-MM-DD UTC format OR a number (number of minutes ago).",
-    )
-
-
-def _add_output_format_arg(parser):
-    parser.add_argument(
-        "-f",
-        "--format",
-        dest="format",
-        action="store",
-        choices=list(OutputFormat()),
-        default=OutputFormat.JSON,
-        help="The format used for outputting events.",
-    )
-
-
-def _add_incremental_arg(parser):
-    parser.add_argument(
-        "-i",
-        "--incremental",
-        dest="is_incremental",
-        action="store_true",
-        help="Only get events that were not previously retrieved.",
     )
 
 
@@ -71,9 +57,7 @@ def _add_exposure_types_arg(parser):
         "--types",
         nargs="*",
         action="store",
-        dest="exposure_types",
+        dest=SearchArguments.EXPOSURE_TYPES,
         choices=list(ExposureType()),
         help="To limit extracted events to those with given exposure types.",
     )
-
-